refactor(blockscout): single source of truth for supported chains
Move the chainId -> instance map to a client-safe chains.ts shared by the
plugin definition and the step runtime. The Chain selector's allowedChainIds
now derives from the same map instead of a hand-maintained list, so adding a
chain is genuinely a one-line edit and the dropdown can't drift from the
instances the runtime actually supports. (The map can't be imported from
blockscout-core directly because that file is server-only and index.ts is
client-loaded.)
